Add asymmetric width parameters
This adds the possibility to configure the left and the right slope in a
different way as well as the distances to the wedge on left and right
side.

The resulting object should not change due to this commit. All
parameters are configured in a way that they are equivalent to the old
symmetric version.